Title: 컴퓨터와 함께하는 숫자 게임 2, Time: -1ms, Memory: -1MB, Status: Runtime Error - CodeTree

+++ 231115/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.java
@@ -25,7 +25,7 @@ public class Main {
     }
     public static void main(String[] args) throws IOException{
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-        m = Integer.parseInt(br.readLine());
+        m = Long.parseLong(br.readLine());
         StringTokenizer st = new StringTokenizer(br.readLine());
         long a = Integer.parseInt(st.nextToken());
         long b = Integer.parseInt(st.nextToken());
